Deduplicate write_pmp
Deduplicating write_pmp is not only a best practice for software engineers but also enhances the coverage of symbolic execution.
